+++ tests/test_core.py
@@ -8,6 +8,7 @@ import pandas as pd
 import pytest
 import xarray as xr
 
+import intake_esm
 from intake_esm.core import _get_subset, _normalize_query, _unique
 
 here = os.path.abspath(os.path.dirname(__file__))
@@ -114,7 +115,7 @@ def test_load_esmcol_remote(zarr_aws_cesmle_col):
 @pytest.mark.parametrize('decode_times', [True, False])
 def test_getitem(sample_cmip6_col, key, decode_times):
     x = sample_cmip6_col[key]
-    assert isinstance(x, intake.catalog.local.LocalCatalogEntry)
+    assert isinstance(x, intake_esm.source.ESMGroupDataSource)
     ds = x(cdf_kwargs={'chunks': {}, 'decode_times': decode_times}).to_dask()
     assert isinstance(ds, xr.Dataset)
     assert set(x.df['member_id']) == set(ds['member_id'].values)
@@ -141,6 +142,15 @@ def test_contains(key, expected):
     col = intake.open_esm_datastore(cdf_col_sample_cmip6)
     actual = key in col
     assert actual == expected
+
+
+def test_df_property():
+    col = intake.open_esm_datastore(catalog_dict_records)
+    assert len(col.df) == 5
+    col.df = col.df.iloc[0:2, :]
+    assert isinstance(col.df, pd.DataFrame)
+    assert len(col) == 1
+    assert len(col.df) == 2
 
 
 def test_serialize_to_json():
@@ -190,19 +200,16 @@ def test_empty_queries(sample_cmip6_col):
 
 
 @pytest.mark.parametrize(
-    'esmcol_path, query, kwargs',
-    [
-        (zarr_col_pangeo_cmip6, zarr_query, {}),
-        (cdf_col_sample_cmip6, cdf_query, {'chunks': {'time': 1}}),
-    ],
+    'esmcol_path, query', [(zarr_col_pangeo_cmip6, zarr_query), (cdf_col_sample_cmip6, cdf_query)]
 )
-def test_to_dataset_dict(esmcol_path, query, kwargs):
+def test_to_dataset_dict(esmcol_path, query):
     col = intake.open_esm_datastore(esmcol_path)
     cat = col.search(**query)
-    if kwargs:
-        _, ds = cat.to_dataset_dict(zarr_kwargs={'consolidated': True}, cdf_kwargs=kwargs).popitem()
-    else:
-        _, ds = cat.to_dataset_dict().popitem()
+    _, ds = cat.to_dataset_dict(
+        zarr_kwargs={'consolidated': True},
+        cdf_kwargs={'chunks': {'time': 1}},
+        storage_options={'token': 'anon'},
+    ).popitem()
     assert 'member_id' in ds.dims
     assert len(ds.__dask_keys__()) > 0
     assert ds.time.encoding
@@ -221,19 +228,20 @@ def test_to_dataset_dict_aggfalse(esmcol_path, query):
 
 
 @pytest.mark.parametrize(
-    'esmcol_path, query, kwargs',
-    [
-        (zarr_col_pangeo_cmip6, zarr_query, {}),
-        (cdf_col_sample_cmip6, cdf_query, {'chunks': {'time': 1}}),
-    ],
+    'esmcol_path, query', [(zarr_col_pangeo_cmip6, zarr_query), (cdf_col_sample_cmip6, cdf_query)],
 )
-def test_to_dataset_dict_w_preprocess(esmcol_path, query, kwargs):
+def test_to_dataset_dict_w_preprocess(esmcol_path, query):
     def rename_coords(ds):
         return ds.rename({'lon': 'longitude', 'lat': 'latitude'})
 
     col = intake.open_esm_datastore(esmcol_path)
     col_sub = col.search(**query)
-    dsets = col_sub.to_dataset_dict(zarr_kwargs={'consolidated': True}, preprocess=rename_coords)
+    dsets = col_sub.to_dataset_dict(
+        zarr_kwargs={'consolidated': True},
+        cdf_kwargs={'chunks': {'time': 1}},
+        preprocess=rename_coords,
+        storage_options={'token': 'anon'},
+    )
     _, ds = dsets.popitem()
     assert 'latitude' in ds.dims
     assert 'longitude' in ds.dims
@@ -251,7 +259,9 @@ def test_to_dataset_dict_w_cmip6preprocessing(pangeo_cmip6_col):
         member_id='r1i1p1f1',
     )
     _, ds = cat.to_dataset_dict(
-        zarr_kwargs={'consolidated': True, 'decode_times': False}, preprocess=combined_preprocessing
+        zarr_kwargs={'consolidated': True, 'decode_times': False},
+        preprocess=combined_preprocessing,
+        storage_options={'token': 'anon'},
     ).popitem()
     assert isinstance(ds, xr.Dataset)
 
@@ -262,10 +272,14 @@ def test_to_dataset_dict_w_cmip6preprocessing(pangeo_cmip6_col):
 def test_to_dataset_dict_nocache(esmcol_path, query):
     col = intake.open_esm_datastore(esmcol_path)
     cat = col.search(**query)
-    _, ds = cat.to_dataset_dict(zarr_kwargs={'consolidated': True}).popitem()
+    _, ds = cat.to_dataset_dict(
+        zarr_kwargs={'consolidated': True}, storage_options={'token': 'anon'}
+    ).popitem()
     id1 = id(ds)
     cat = col.search(**query)
-    _, ds = cat.to_dataset_dict(zarr_kwargs={'consolidated': True}).popitem()
+    _, ds = cat.to_dataset_dict(
+        zarr_kwargs={'consolidated': True}, storage_options={'token': 'anon'}
+    ).popitem()
     assert id1 != id(ds)
 
 
@@ -319,6 +333,7 @@ def test_to_dataset_dict_s3(zarr_aws_cesmle_col):
 def test_read_catalog_dict():
     col = intake.open_esm_datastore(catalog_dict_records)
     assert isinstance(col.df, pd.DataFrame)
+    assert col.catalog_file is None
 
 
 params = [